+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Numbers inputted into a number variable are validated on defined min/max and if it is a number
+- Fragment insert modal
 
 ### Fixed
 - Fixed woodpecker builds crashing on syntax errors
@@ -34,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [8.0.1] - 2023-06-15
 ### Fixed
 - Change problematic type in citation that made it to break with new ember
+
 
 ## [8.0.0] - 2023-06-15
 ### Fixed